Revert modals.js typescript conversion. Remove click event typings and coerce other types to minimize number of files changed.
