--- components/widgets/pulsating_dot/index.ts
@@ -1,4 +1,0 @@
-// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
-// See LICENSE.txt for license information.
-
-export {default} from './pulsating_dot';
