+++ components/widgets/pulsating_dot/index.tsx
